+++ src/main/java/com/sumirelabs/pulsar/light/engine/ScalarSkyEngine.java
@@ -55,7 +55,7 @@ public class ScalarSkyEngine extends PulsarEngine {
 
     @Override
     protected boolean canUseChunk(final Chunk chunk) {
-        return ((PulsarChunk) chunk).pulsar$isLightReady();
+        return ((PulsarChunk) chunk).pulsar$isLightUsable();
     }
 
     @Override